download-qrcode: new qrcode component and download
