+++ frontend/package.json
@@ -39,16 +39,17 @@
     "dotenv": "^16.3.1",
     "eslint-config-prettier": "^9.0.0",
     "floating-vue": "^2.0.0-beta.24",
+    "html2canvas": "^1.4.1",
     "leaflet": "^1.9.4",
     "nuxt-mail": "^3.1.23",
     "pinia": "^2.1.6",
-    "uuid": "^9.0.1",
     "prettier": "^3.0.3",
     "qrcode": "^1.5.3",
     "qrcode.vue": "^3.4.1",
     "swiper": "^9.3.2",
     "tailwind-scrollbar": "^3.0.4",
     "tailwindcss": "^3.3.2",
+    "uuid": "^9.0.1",
     "v-calendar": "^3.1.1"
   }
 }
